LayoutTests/storage/indexeddb/setversion-blocked-by-versionchange-close-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test open and setVersion are blocked if a VERSION_CHANGE transaction is running, and connection closes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "setversion-blocked-by-versionchange-close.html"
ver = 1
blockedEventFired = false
versionChangeComplete = false
indexedDB.open(dbname)

h1OpenOnSuccess():
h1 = event.target.result
indexedDB.open(dbname)

h2OpenOnSuccess():
h2 = event.target.result
h1.setVersion(String(ver++))
indexedDB.open(dbname)
h2.close()

h1SetVersionOnBlocked():
blockedEventFired = true

h1SetVersionOnSuccess():

transactionOnComplete():
versionChangeComplete = true

h3OpenOnSuccess():
FIXME: blocked should not fire as connection was closed. http://webkit.org/b/71130
FAIL blockedEventFired should be false. Was true.
PASS versionChangeComplete is true
PASS successfullyParsed is true

TEST COMPLETE
